Stage3Wave5の下半分の右側が少し早く出現する問題を解決しました。 原因としては同様にINIT_ARG()関数の戻り値がdoubleではなくintだったためです。 Stage3Wave5について、少しバランス調整も行いました。

+++ program/Character/EnemyCharacter/ZakoCharacter/ZkChrStg3Wv5BLR.h
@@ -28,7 +28,7 @@ private:
 	static const double DRAW_EXTRATE;
 
 	static int INIT_POS_X(Stg3WAVE5BLR given_lr);
-	static int INIT_ARG(Stg3WAVE5BLR given_lr);
+	static double INIT_ARG(Stg3WAVE5BLR given_lr);
 	
 public:
 	ZkChrStg3Wv5BLR(CharacterID given_id, Stg3WAVE5BLR lr);
